Update README.md
I started in the 6th grade, shooting a ball in a hoop in the playground
the fun started to increase once you liked winning 
Then competing with everyone around you to become the best
Shooting from the left and right
Then the need to have fancy dribbles allows it be even more chill

My basketball ambitions picked up once i was on the church team
Every practice working on form shooting, making my jump shot better than before
I had the height, i have the speed, i have the youth, i have everything in my arresnal to become a great basketball player
Yet, over my years,i became stagnant until recently. My jump shot wasn't very consisent from outside. I only made the right hand floater layup and occasionally used my footwork to beat out using fakes. But i wasn't able to sell it
I learned the jab step but not to the point where it should be an instant move. I have to learn to body and work around
I don't have a floater nor any fancy dribbling
I enjoy playing defense and hustling because im good at it
I can't shut down players, but i want to be a player who prides himself on defense because anyone can do it
no one has no unfair advantage on the amount of pressure one person can apply 
Also i enjoy playing team ball and taking good selections

I have to be more confident in my shot also understand not to let the pressure get into my head
During closing points of close games, I begin to make careless mistakes which cost us the game or turnovers. 

I panic and the fear in my eyes makes it highly obvious im scared as shit right now. I have to learn to take it calmly and play through the possesion. I get caught up in the moment feeling like the weight is upon me when it is a team sports, everyone on the team trusts you to handle the ball, make a play if required and play with 100% effort. 

During our loss in the KOC finals, I went exteremely hard to ensure I played amazing defense. I gave effort from the beginning to the end. Granted, it was first actual minutes and it was due to the fact our starter Anthony cramped up and it was between Aaron and myself to step up and finsih the game. 

This really allowed me to get important possesions and finishing the game with more positives than negative. Regardless I shouldn't have air balled my one mid range shot. I have to be able to spot of players during in anbound play. I have to understand how to clear the court for kp or blesson to take the charge. Also, i have to speak up if i see a bad play coming and tell them to make a quick decision. I remeber junior year, i was in the middle of the game with nevin and he was going to pull a shitty 3 when he clearly wasn't a spot on shooter, so i advised him to drive and take the contact. He made the layup, it was an easily looked over veteran prescence.

When Blesson is on the court, he makes it exteremly easy to be comfortable playing because he slows down the game for the rest of the guys. He takes all the pressure off you to score and he steps up on defense. When someone makes a mistake, he is very quick to point out the trouble and make you understand what is wrong and right. Utimately the one thing that stuck with me in ball and running would be "Take it one game at a time, were not in the finals, we have to worry about the cleaning up our act right now, so later when the playoffs and finals come , we are ready." He really is a player who walks it like he talks it.

In my quest to support the team, i raelized we try to engage and fix anything we don't like as quickly as possible. We try to find solutions quickly, and think through the game plan. The pursuit to the ultimate goal is never considered a far too hard task to achieve. With some hardwork, practice and desire to change your own game for the desire of the team winnning everyting allows you to take your game to the next level. 

Every practice I would look up to the chance to guard Blesson or KP, i take it as a challenge to prove yourself and understand you are capable of locking down some of the best talent in Mallu Ball. KP can make the case for top 5 arguable best brown player in the region, blesson is a top 10 player and one of the smartest veteran players i ever spoken to. 

The team grew strong under the faith of God and under his guidance. We pray before every game and after every game for all the blessings God has placed upon the team and carrying us through the journey with his shield. Keeping us in the path to the promised land. When we pray, we all huddle and it becomes a small spiritual adoration where we know whats about to happen and we will go out there with God on our side. 

One thing the team and myself can really learn from is Adversity. For the past couple years since our last actual ring in KOC in 2016, we have been dealing with getting outplayed, out hussled, and out smarted in the playoffs and finals. Making it to the finals and losing every year, has its benefits and its fall. Sometimes, we don't even make the playoffs. When our team is down, we don't know how to play. WE are extermely lousy with the ball, everyone wants to play hero ball, every player out there is uneasy. Its sloopy basketball to say in short. We have to learn to play with a controled temper and pace and move the ball. Ultimately, as the other team is up, its our job to lock up on defense with extra effort all around, then on offense moving the ball around a couple possesions before pulling the trigger. It can't just be one person giving all the effort because the team will lose at that point. We have to be able to find a way to get the ball in the bucket. 

I trust the process that my potential can be exteremly good if i put in the work and continue to get more game exposure. One thing, Donvon Mitchell did stress is the ability that he is allowed to make mistakes on the court, study the film and be mroe prepared the follwoing games. 

I want to  be the best version i am allowed to become on top of obtaining a high paying job, graduating college, developing into an all around player on and off the court.